ENG-2539 prepare to upgrade Apollo tooling for `yarn schema:*` scripts
- simpler `serverOptions` logic
- "schemaTag" => "graphVariant" wherever possible
- support both future `APOLLO_KEY` and deprecated `ENGINE_API_KEY`
- FIXMEs for changes to be made after server-core Apollo upgrades

these changes have been smoke-tested with both `event_service` and `stitch_service`.  `event_service` will be the guinea pig in terms of an across-the-board version upgrade for Apollo + TypeORM, in anticipation of the formal 'server-core 2.0' effort.  `stitch_service` will take on the latest server-core *without* any version upgrades to Apollo (a different guinea pig scenario).

we are **not** taking on the following across-the-board version upgrades (which were adopted by `event_service` and smoke-tested in `stitch_service`)

```
"@apollo/federation": "^0.16.2",
"@apollo/gateway": "^0.14.1", // or we'd need to switch to Node 12
"@types/express-serve-static-core": "^4.17.7",
"@types/express": "^4.17.4",
"apollo-server-express": "^2.14.1",
"apollo-server-testing": "^2.14.1",
"apollo-server": "^2.14.1",
"apollo": "^2.27.0",
"typeorm": "~0.2.25"
```