CORE fixing issues with sanitizeUrlString and its sibling
the way it was built, it would *always* convert the URL to 'https:'.  i'm shocked that this didn't bite us before.  preserving the protocol during the sanitization process turns out to require some RegExp trickery, mostly because Node 10.  i made the behavior conform to the prior spec per the Test Suite, then pushed it a bit farther to see what would happen